Update README.md
In this SQL file uploaded to GitHub, I've embarked on a journey through various database operations, showcasing my grasp of SQL queries. The exploration begins by using the DESCRIBE statement to shed light on the structure of the "customer" table, unveiling its columns, data types, and constraints. This initial step offers a quick reference for anyone delving into the dataset.

Moving forward, a series of SELECT statements unfolds, revealing the contents of the "customer" and "billing" tables. These queries not only display the data within these tables but also serve as a testament to my ability to navigate and extract specific information from a relational database.

As the queries progress, I delve into more intricate operations. The selective extraction of columns from the "customer" table using SELECT showcases my proficiency in tailoring data output. The utilization of the DISTINCT keyword demonstrates my understanding of data deduplication, particularly when identifying unique email addresses in the "customer" table.

The application of aggregate functions becomes evident as I calculate the sum of the "amount_due" in the "billing" table, showcasing my knack for performing mathematical operations on dataset elements.

Furthermore, the incorporation of the WHERE clause underscores my skill in filtering data based on specific conditions. I retrieve records of customers with the first name "Tore" and bills with amounts exceeding 5000, showcasing my ability to extract relevant information from large datasets.

Queries involving the LIKE and IN operators demonstrate my understanding of pattern matching and multiple condition checks. The use of sorting and limiting techniques with ORDER BY and LIMIT illustrates my adeptness in organizing and restricting result sets for clearer insights.

Each query is accompanied by detailed comments and explanations, reflecting my commitment to effective code documentation. This file, thus, stands as a comprehensive exhibit of my SQL proficiency, serving both as a personal reference and a showcase for potential collaborators or employers on GitHub.